Split Setting Activity to Activity & Fragment
